# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Career Progression Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /career-paths:
    get:
      operationId: listCareerPaths
      tags: [careerPaths]
      summary: List all career paths
      responses:
        "200":
          description: List of career paths
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CareerPath"
    post:
      operationId: createCareerPath
      tags: [careerPaths]
      summary: Create a career path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCareerPathBody"
      responses:
        "201":
          description: Created career path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerPath"
  /career-paths/{id}:
    get:
      operationId: getCareerPath
      tags: [careerPaths]
      summary: Get a career path by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Career path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerPath"
        "404":
          description: Not found
  /roles:
    get:
      operationId: listRoles
      tags: [roles]
      summary: List roles, optionally filtered by career path
      parameters:
        - name: careerPathId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
    post:
      operationId: createRole
      tags: [roles]
      summary: Create a role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleBody"
      responses:
        "201":
          description: Created role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
  /roles/{id}:
    get:
      operationId: getRole
      tags: [roles]
      summary: Get a role with its full job spec and competencies
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Role with competencies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleWithCompetencies"
        "404":
          description: Not found
  /competencies:
    get:
      operationId: listCompetencies
      tags: [competencies]
      summary: List competencies for a role
      parameters:
        - name: roleId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of competencies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Competency"
    post:
      operationId: createCompetency
      tags: [competencies]
      summary: Create a competency for a role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompetencyBody"
      responses:
        "201":
          description: Created competency
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competency"
...
  /evidence:
    get:
      operationId: listEvidence
      tags: [evidence]
      summary: List evidence entries for a user and role
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
        - name: roleId
          in: query
          required: false
          schema:
            type: integer
        - name: competencyId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of evidence entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Evidence"
    post:
      operationId: createEvidence
      tags: [evidence]
      summary: Add an evidence entry for a competency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvidenceBody"
      responses:
        "201":
          description: Created evidence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evidence"
  /evidence/{id}:
    put:
      operationId: updateEvidence
      tags: [evidence]
      summary: Update an evidence entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEvidenceBody"
      responses:
        "200":
          description: Updated evidence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evidence"
    delete:
      operationId: deleteEvidence
      tags: [evidence]
      summary: Delete an evidence entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: managerId
          in: query
          required: false
          schema:
            type: integer
        - name: role
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    put:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /manager/team:
    get:
      operationId: getManagerTeam
      tags: [manager]
      summary: Get all team members for a manager
      parameters:
        - name: managerId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManagerTeamMember"
  /manager/dashboard-stats:
    get:
      operationId: getManagerDashboardStats
      tags: [manager]
      summary: Get probation dashboard stats for a manager
      parameters:
        - name: managerId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagerDashboardStats"
components:
  schemas:
...
    CareerPath:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - createdAt
    CreateCareerPathBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required:
        - name
        - description
    Role:
      type: object
      properties:
        id:
          type: integer
        careerPathId:
          type: integer
        title:
          type: string
        level:
          type: integer
        jobSpec:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - careerPathId
        - title
        - level
        - jobSpec
        - createdAt
    RoleWithCompetencies:
      type: object
      properties:
        id:
          type: integer
        careerPathId:
          type: integer
        title:
          type: string
        level:
          type: integer
        jobSpec:
          type: string
        createdAt:
          type: string
          format: date-time
        competencies:
          type: array
          items:
            $ref: "#/components/schemas/Competency"
      required:
        - id
        - careerPathId
        - title
        - level
        - jobSpec
        - createdAt
        - competencies
    CreateRoleBody:
      type: object
      properties:
        careerPathId:
          type: integer
        title:
          type: string
        level:
          type: integer
        jobSpec:
          type: string
      required:
        - careerPathId
        - title
        - level
        - jobSpec
    Competency:
      type: object
      properties:
        id:
          type: integer
        roleId:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - roleId
        - name
        - description
        - category
        - createdAt
    CreateCompetencyBody:
      type: object
      properties:
        roleId:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
      required:
        - roleId
        - name
        - description
        - category
...
    Evidence:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        competencyId:
          type: integer
        roleId:
          type: integer
        title:
          type: string
        description:
          type: string
        rating:
          type: string
          enum: [red, amber, green]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - competencyId
        - roleId
        - title
        - description
        - rating
        - createdAt
        - updatedAt
    CreateEvidenceBody:
      type: object
      properties:
        userId:
          type: integer
        competencyId:
          type: integer
        roleId:
          type: integer
        title:
          type: string
        description:
          type: string
        rating:
          type: string
          enum: [red, amber, green]
      required:
        - userId
        - competencyId
        - roleId
        - title
        - description
        - rating
    UpdateEvidenceBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        rating:
          type: string
          enum: [red, amber, green]
      required:
        - title
        - description
        - rating
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        roles:
          type: array
          items:
            type: string
        managerId:
          type: ["integer", "null"]
        department:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        probationStatus:
          type: ["string", "null"]
        targetRoleId:
          type: ["integer", "null"]
        isActive:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - roles
        - isActive
        - createdAt
        - updatedAt
    UserInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
        managerId:
          type: integer
        department:
          type: string
        jobTitle:
          type: string
        startDate:
          type: string
        probationStatus:
          type: string
        targetRoleId:
          type: integer
        isActive:
          type: string
      required:
        - name
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
        managerId:
          type: ["integer", "null"]
        department:
          type: string
        jobTitle:
          type: string
        startDate:
          type: string
        probationStatus:
          type: string
        targetRoleId:
          type: ["integer", "null"]
        isActive:
          type: string
    ManagerTeamMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        probationStatus:
          type: ["string", "null"]
        isActive:
          type: string
        latestReviewPeriod:
          type: ["string", "null"]
        latestReviewPublishedAt:
          type: ["string", "null"]
          format: date-time
        reviewCount:
          type: integer
      required:
        - id
        - name
        - isActive
        - reviewCount
    ManagerDashboardStats:
      type: object
      properties:
        totalTeam:
          type: integer
        inProbation:
          type: integer
        pendingReviews:
          type: integer
        publishedReviews:
          type: integer
        needingAttention:
          type: integer
      required:
        - totalTeam
        - inProbation
        - pendingReviews
        - publishedReviews
        - needingAttention