# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Career Progression Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  roles: string[];
  /** @nullable */
  managerId?: number | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  jobTitle?: string | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  probationStatus?: string | null;
  /** @nullable */
  targetRoleId?: number | null;
  isActive: string;
  createdAt: Date;
  updatedAt: Date;
}