# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Career Progression Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  name?: string;
  email?: string;
  roles?: string[];
  /** @nullable */
  managerId?: number | null;
  department?: string;
  jobTitle?: string;
  startDate?: string;
  probationStatus?: string;
  /** @nullable */
  targetRoleId?: number | null;
  isActive?: string;
}